♻️ Extract @rallly/api-core package (RAL-1172)
Move the framework-agnostic, Prisma-touching poll mutation/query and API
key crypto modules out of apps/web into a new @rallly/api-core workspace
package so the upcoming standalone apps/api can consume them alongside
web's tRPC.

URL construction (adminUrl/inviteUrl/inviteLink) lives in callers, not
in api-core, so the package has no env var coupling.
