+++ apps/web/package.json
@@ -36,6 +36,7 @@
     "@radix-ui/react-select": "^2.2.2",
     "@radix-ui/react-slot": "^1.1.2",
     "@radix-ui/react-switch": "^1.0.2",
+    "@rallly/api-core": "workspace:*",
     "@rallly/billing": "workspace:*",
     "@rallly/database": "workspace:*",
     "@rallly/emails": "workspace:*",
