+++ apps/web/src/app/api/private/[...route]/route.ts
@@ -1,4 +1,9 @@
 import { RedisStore } from "@hono-rate-limiter/redis";
+import {
+  getPollParticipants,
+  getPollResults,
+} from "@rallly/api-core/polls/data";
+import { createPoll, deletePoll } from "@rallly/api-core/polls/mutations";
 import { prisma } from "@rallly/database";
 import { absoluteUrl, shortUrl } from "@rallly/utils/absolute-url";
 import { Scalar } from "@scalar/hono-api-reference";
@@ -11,8 +16,6 @@ import {
   validator,
 } from "hono-openapi";
 import { rateLimiter } from "hono-rate-limiter";
-import { getPollParticipants, getPollResults } from "@/features/poll/data";
-import { createPoll, deletePoll } from "@/features/poll/mutations";
 import { redis } from "@/lib/kv";
 import {
   createPollInputSchema,
@@ -210,8 +213,8 @@ app.post(
             startTime: option.startTime.toISOString(),
             duration: option.duration,
           })),
-          adminUrl: poll.adminUrl,
-          inviteUrl: poll.inviteUrl,
+          adminUrl: absoluteUrl(`/poll/${poll.id}`),
+          inviteUrl: shortUrl(`/invite/${poll.id}`),
         },
       });
     }
@@ -301,8 +304,8 @@ app.post(
           startTime: option.startTime.toISOString(),
           duration: option.duration,
         })),
-        adminUrl: poll.adminUrl,
-        inviteUrl: poll.inviteUrl,
+        adminUrl: absoluteUrl(`/poll/${poll.id}`),
+        inviteUrl: shortUrl(`/invite/${poll.id}`),
       },
     });
   },
