+++ apps/web/src/app/api/private/utils/api-key.ts
@@ -1,8 +1,11 @@
 import crypto from "node:crypto";
+import {
+  extractApiKeyPrefix,
+  verifyApiKey,
+} from "@rallly/api-core/auth/api-key";
 import { prisma } from "@rallly/database";
 import { bearerAuth } from "hono/bearer-auth";
 import { after } from "next/server";
-import { extractApiKeyPrefix, verifyApiKey } from "@/features/developer/utils";
 
 export const spaceApiKeyAuth = bearerAuth({
   verifyToken: async (rawKey, c) => {
