+++ packages/api-core/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@rallly/api-core",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "test:unit": "vitest run",
+    "type-check": "tsc --noEmit"
+  },
+  "exports": {
+    "./*": "./src/*.ts"
+  },
+  "dependencies": {
+    "@rallly/database": "workspace:*",
+    "@rallly/logger": "workspace:*",
+    "@rallly/utils": "workspace:*"
+  },
+  "devDependencies": {
+    "@rallly/tsconfig": "workspace:*",
+    "@types/node": "^24.1.0",
+    "vitest": "^4.0.16"
+  }
+}
