+++ packages/api-core/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@rallly/tsconfig/node.json",
+  "compilerOptions": {
+    "types": ["vitest/globals", "node"],
+    "lib": ["es2020", "dom"]
+  },
+  "include": ["**/*.ts"]
+}
